Math (#14)
* MathJax support

There was no support for displaying math written in Latex. This is now added with the help of MathJax.

MathJax uses an external Javascript file and as not to add unnecessary load the support for math is only enabled if the 'math' param is set to true.

* MathJax support

There was no support for displaying math written in Latex. This is now added with the help of MathJax.

MathJax uses an external Javascript file and as not to add unnecessary load the support for math is only enabled if the 'math' param is set to true.

* Auto stash before merge of "math" and "origin/math"

* Delete unnecessary option

* Actually enable math for demo page

* Have demo page be consistent