Check if commentoUrl is set before using it. (#198)
* Check if commentoUrl is set before using it.

When a variable does not exist it evaluates to `nil`, and so the
check `(not (eq .Site.Params.commentoUrl ""))` evaluates to true rather
than false because `nil` is not "".  Before this change, unless you
explicitly set Params.commentoUrl to "", the commento HTML block would
be added, and this fails because /js/commento.js does not exist.  This
change ensures that the commento HTML block is only added whe
Params.commentoUrl exists and is set to a non-empty string.

* Check if DisqusShortname is set before using it.