Fix bad syntax highlighting color schemes
Some pygment color schemes (like `bw`, which we use as default in the
exampleSite) do not set a foreground color explicitly, which makes those
themes look bad/illegible when using the dark mode of hugo-coder. Most
such themes are light themes (it seems after a quick research), which
explains this phenomenon.

This commit adds a fix for such color schemes by setting the default
foreground color to the one used in the light hugo-coder colorscheme.
It is overridden by proper syntax highlighting schemes.