Update Plausible script (#738)
### Prerequisites

Put an `x` into the box(es) that apply:

- [ ] This pull request fixes a bug.
- [x] This pull request adds a feature.
- [ ] This pull request introduces breaking change.

### Description

Fixes an issue with the Plausible script not loading. As per the
[docs](https://plausible.io/docs/plausible-script), it should be
`script.js` and not `plausible.js`.

**Update:** After further testing (CSP appeared to be the issue), it
looks like it also works with `plausible.js`, probably through backwards
compatibility. If needed, for consistency, this can still be updated to
`script.js`.

### Issues Resolved

Related to #388. It looks like they updated it from `plausible.js` to
`script.js`.

### Checklist

Put an `x` into the box(es) that apply:

#### General

- [x] Describe what changes are being made
- [x] Explain why and how the changes were necessary and implemented
respectively
- [x] Reference issue with `#<ISSUE_NO>` if applicable

#### Resources

- [ ] If you have changed any SCSS code, run `make release` to
regenerate all CSS files

#### Contributors

- [x] Add yourself to `CONTRIBUTORS.md` if you aren't on it already