Add en & pt-br translation of 'Hugo Authors' authors
You can add a slug to the taxonomy frontmatter, but it won't be used
unless you add a section

    [permalinks]
    authors = "/authors/:slug"

to your config.toml, and even then "authors" isn't translated when
you switch language. You can use "url" in the taxonomy frontmatter
instead to specify "autores/" and then the slug to get the url
translated, without needing to change site config.