Apply larger h1 size to all <header> H1s
The current CSS only applies it to H1s inside article tags, but many pages are not articles and therefore don't get the resize.  This leads to different pages having differing heights of header content.

This also has to actually add header tags to headers on some partials that don't use the semantic tags right now.