Make the navigation menu keyboard-accessible (#833)
### Prerequisites

Put an `x` into the box(es) that apply:

- [x] This pull request fixes a bug.
- [x] This pull request adds a feature.
- [ ] This pull request introduces breaking change.

### Description

This PR makes the navigation hamburger menu accessible via a keyboard.
Since elements with `display: none` are not focusable, I decided instead
to reset the `display` property of the element to its default value. The
checkbox is now hidden in a different way by positioning it outside of
the viewport and making it transparent. This way, the checkbox is not be
visible, but is still focusable.

The checkbox can be focused by tabbing on a keyboard, and once the
checkbox is focused, pressing space will open or close the menu. Since
the checkbox itself is invisible, I have added a style to the label
element (the hamburger menu icon) that indicates the focus.

### Issues Resolved

#832

### Checklist

Put an `x` into the box(es) that apply:

#### General

- [x] Describe what changes are being made
- [x] Explain why and how the changes were necessary and implemented
respectively
- [x] Reference issue with `#<ISSUE_NO>` if applicable

#### Resources

- [x] If you have changed any SCSS code, run `make release` to
regenerate all CSS files

#### Contributors

- [x] Add yourself to `CONTRIBUTORS.md` if you aren't on it already