Swap from ForkAwesome 1.2.0 to FontAwesome 6.5.1. (#882)
### Prerequisites

Put an `x` into the box(es) that apply:

- [x] This pull request fixes a bug.
- [x] This pull request adds a feature.
- [x] This pull request introduces breaking change.

### Description

This change allows FontAwesome 6.5.1 to be used as a drop-in replacement
for ForkAwesome 1.2.0.

The easiest way to see that my change has worked practically-invisibly
is that the home page of the demo site now shows Medium with its new
three-circular-blobs logo instead of the old capital-M logo, despite the
fact I have not changed anything about the hugo configuration for the
demo site.

Points of note here:

- Using FontAwesome in SCSS is different. Therefore the ONLY breaking
change here is for people who have made changes to the theme and written
"@include fa-icon" in their SCSS. Everything that can happen in content
(e.g. class names) are translated by the v4-shims.
- All files here are from https://fontawesome.com/download ('Free for
Web').
  - The contents of the /scss folder are in assets/scss/font-awesome.
- The contents of /webfonts are in static/fonts (except
fa-v4compatibility because that's not needed - v4-shims does all the
work we need).
- /LICENSE.txt is added to both directories to ensure it is clear both
are part of the FA-licensed code.
- In ForkAwesome, the path to the font was constructed in _path.css
suffixing $fa-version, so the preload needed to have the version at the
end. Since the default behaviour of FontAwesome is to not include a
version suffix, the preloads no longer have one either. This is a
necessary change because the HTML spec (section 4.6.6.19) explicitly
says that the entire URL [i.e. including query and fragment] is in the
key used to determine whether an element in the preload cache can be
consumed.

### Issues Resolved

#858 

### Checklist

Put an `x` into the box(es) that apply:

#### General

- [x] Describe what changes are being made
- [x] Explain why and how the changes were necessary and implemented
respectively
- [x] Reference issue with `#<ISSUE_NO>` if applicable

#### Resources

- [x] If you have changed any SCSS code, run `make release` to
regenerate all CSS files

#### Contributors

- [x] Add yourself to `CONTRIBUTORS.md` if you aren't on it already

Co-authored-by: Luiz F. A. de Prá <luizdepra@users.noreply.github.com>