Update docs for `docs/analytics/onedollarstats.md` (#948)
This is a followup to #944. 

The content security policy had to be updated for this to work.

I worked with the @ drizzle to get this working. Discord thread is
available
[here](https://discord.com/channels/1043890932593987624/1315096480041799720).

Screenshot as a verification for posterity:

<img width="971" alt="Screenshot 2024-12-21 at 4 08 00 PM"
src="https://github.com/user-attachments/assets/4c9a1a60-9f3a-4acc-a887-be1490e1add5"
/>

Co-authored-by: Luiz F. A. de Prá <luizdepra@users.noreply.github.com>