Update deploy.yml missing sass still
 log confirms a few key points:

ruby-sass Removal:
It's still trying to remove ruby-sass and is doing so successfully.
dart-sass installed successfully:
The dart-sass install command is running without error.
Exit Code 1:
The workflow is still exiting with code 1, which means there is still an error.
The Problem:

The key thing that is missing from these logs, is the output of the which sass command. We need to see where the sass executable is being installed.